Added some checks when saving the files, in case the file to check that is being upgraded on a different thread. Fix main_brid

+++ master/buildbot/status/build.py
@@ -462,6 +462,9 @@ class BuildStatus(styles.Versioned, properties.PropertiesMixin):
             # leftover from 0.5.0, which stored builds in directories
             shutil.rmtree(filename, ignore_errors=True)
         tmpfilename = filename + ".tmp"
+
+        if os.path.exists(tmpfilename):
+            return
         try:
             with open(tmpfilename, "wb") as f:
                 dump(self, f, -1)
@@ -472,7 +475,9 @@ class BuildStatus(styles.Versioned, properties.PropertiesMixin):
                 # general (non-windows) case
                 if os.path.exists(filename):
                     os.unlink(filename)
-            os.rename(tmpfilename, filename)
+
+            if os.path.exists(tmpfilename):
+                os.rename(tmpfilename, filename)
         except:
             log.msg("unable to save build %s-#%d" % (self.builder.name,
                                                      self.number))
